+++ cmd/veneur-emit/main.go
@@ -13,6 +13,8 @@ import (
 	"syscall"
 	"time"
 
+	"golang.org/x/net/context"
+
 	"fmt"
 	"strconv"
 
@@ -24,6 +26,7 @@ import (
 	"github.com/stripe/veneur/v14/protocol"
 	"github.com/stripe/veneur/v14/ssf"
 	"github.com/stripe/veneur/v14/trace"
+	"google.golang.org/grpc"
 )
 
 type EmitMode uint
@@ -33,6 +36,7 @@ type Flags struct {
 	Mode      string
 	Debug     bool
 	Command   bool
+	Proxy     string
 	ExtraArgs []string
 
 	Name   string
@@ -42,6 +46,7 @@ type Flags struct {
 	Set    string
 	Tag    string
 	ToSSF  bool
+	ToGrpc bool
 
 	Event struct {
 		Title      string
@@ -176,15 +181,24 @@ func Main(args []string) int {
 
 	validateFlagCombinations(passedFlags, flagStruct.ExtraArgs)
 
-	addr, netAddr, err := destination(flagStruct.HostPort, flagStruct.ToSSF)
+	addr, netAddr, err := destination(flagStruct.HostPort)
+	proxy, proxyAddr, err := destination(flagStruct.Proxy)
 	if err != nil {
 		logrus.WithError(err).Error("Error getting destination address.")
 		return 1
 	}
-	logrus.WithField("net", netAddr.Network()).
-		WithField("addr", netAddr.String()).
-		WithField("ssf", flagStruct.ToSSF).
+
+	logrus.WithField("addr", addr).
+		WithField("net", netAddr.Network()).
+		WithField("destination", flagStruct.HostPort).
+		WithField("grcp", flagStruct.ToGrpc).
 		Debugf("destination")
+
+	logrus.WithField("addr", proxy).
+		WithField("net", proxyAddr.Network()).
+		WithField("proxy", flagStruct.Proxy).
+		Debugf("proxy")
+
 
 	if flagStruct.Mode == "event" {
 		if flagStruct.ToSSF {
@@ -241,9 +255,9 @@ func Main(args []string) int {
 		return 1
 	}
 	if span.TraceId != 0 {
-		if !flagStruct.ToSSF {
+		if !flagStruct.ToSSF && !flagStruct.ToGrpc  {
 			logrus.WithField("ssf", flagStruct.ToSSF).
-				Error("Can't use tracing in non-ssf operation: Use -ssf to emit trace spans.")
+				Error("Can't use tracing in non-ssf operation: Use -ssf or -grpc to emit trace spans.")
 			return 1
 		}
 		logrus.WithField("trace_id", span.TraceId).
@@ -260,6 +274,12 @@ func Main(args []string) int {
 		return 1
 	}
 	if flagStruct.ToSSF {
+		if proxyAddr != nil {
+			logrus.WithField("grcp", flagStruct.ToGrpc).
+				Error("Can't use proxy for non grcp protocol: use -grcp with -proxy")
+			return 1
+		}
+
 		client, err := trace.NewClient(addr)
 		if err != nil {
 			logrus.WithError(err).
@@ -273,18 +293,43 @@ func Main(args []string) int {
 			logrus.WithError(err).Error("Could not send SSF span")
 			return 1
 		}
+	} else if flagStruct.ToGrpc {
+		grpcDialOptions := []grpc.DialOption{grpc.WithBlock(), grpc.WithInsecure()}
+		if proxyAddr != nil {
+			grpcDialOptions = append(grpcDialOptions, grpc.WithDialer(func(_ string, timeout time.Duration) (net.Conn, error) {
+				return net.DialTimeout(proxyAddr.Network(), proxyAddr.String(), timeout)
+			}))
+		}
+		conn, err := grpc.Dial(addr, grpcDialOptions...)
+		if err != nil {
+			logrus.WithError(err).Error("Could not dial grpc stats server")
+			return 1
+		}
+		defer conn.Close()
+		client := ssf.NewSSFGRPCClient(conn)
+		_, err = client.SendSpan(context.Background(), span)
+		if err != nil {
+			logrus.WithError(err).Error("Could not send span over grpc")
+			return 1
+		}
 	} else {
-		if netAddr.Network() != "udp" {
+		if proxyAddr != nil {
+			logrus.WithField("grcp", flagStruct.ToGrpc).
+				Error("Can't use proxy for non grcp protocol: use -grcp with -proxy")
+			return 1
+		}
+
+		if netAddr.Network() != "udp" && netAddr.Network() != "tcp" {
 			logrus.WithField("address", addr).
 				WithField("network", netAddr.Network()).
-				Error("hostport must be a UDP address for statsd metrics")
+				Error("hostport must be a UDP or TCP address for statsd metrics")
 			return 1
 		}
 		if len(span.Metrics) == 0 {
 			logrus.Error("No metrics to send. Must pass metric data via at least one of -count, -gauge, -timing, or -set.")
 			return 1
 		}
-		err = sendStatsd(netAddr.String(), span)
+		err = sendStatsd(netAddr, span)
 		if err != nil {
 			logrus.WithError(err).Error("Could not send UDP metrics")
 			return 1
@@ -380,15 +425,24 @@ func tagsFromString(csv string) map[string]string {
 	return tags
 }
 
-func destination(hostport string, useSSF bool) (string, net.Addr, error) {
+func destination(hostport string) (string, net.Addr, error) {
 	if hostport == "" {
 		return "", nil, errors.New("you must specify a valid hostport")
 	}
+
+	hostport, addr, err := resolveHostport(hostport)
+	if err != nil {
+		return "", nil, err
+	}
+	return hostport, addr, nil
+}
+
+func resolveHostport(hostport string) (string, net.Addr, error) {
 	netAddr, err := protocol.ResolveAddr(hostport)
 	if err != nil {
 		// This is fine - we can attempt to treat the
 		// host:port combination as a UDP address:
-		hostport = fmt.Sprintf("udp://%s", hostport)
+		hostport := fmt.Sprintf("udp://%s", hostport)
 		udpAddr, err := protocol.ResolveAddr(hostport)
 		if err != nil {
 			return "", nil, err
@@ -574,10 +628,55 @@ func sendSSF(client *trace.Client, span *ssf.SSFSpan) error {
 	return <-done
 }
 
+// newDatadogTCPWriter is adapted from https://github.com/DataDog/datadog-go/blob/master/statsd/udp.go
+func newDatadogTCPWriter(addr string) (*datadogTCPWriter, error) {
+	tcpAddr, err := net.ResolveTCPAddr("tcp", addr)
+	if err != nil {
+		return nil, err
+	}
+	conn, err := net.DialTCP("tcp", nil, tcpAddr)
+	if err != nil {
+		return nil, err
+	}
+	writer := &datadogTCPWriter{conn: conn}
+	return writer, nil
+}
+
+type datadogTCPWriter struct {
+	conn net.Conn
+}
+
+func (w *datadogTCPWriter) Write(data []byte) (n int, err error) {
+	return w.conn.Write(data)
+}
+
+func (w *datadogTCPWriter) SetWriteTimeout(timeout time.Duration) error {
+	// This is unused in the current implementation
+	return nil
+}
+
+func (w *datadogTCPWriter) Close() error {
+	return w.conn.Close()
+}
+
+
 // sendStatsd sends the metrics gathered in a span to a dogstatsd
 // endpoint.
-func sendStatsd(addr string, span *ssf.SSFSpan) error {
-	client, err := statsd.New(addr)
+func sendStatsd(addr net.Addr, span *ssf.SSFSpan) error {
+	var client *statsd.Client
+	var err error
+	network := addr.Network()
+	switch network {
+	case "udp":
+		client, err = statsd.New(addr.String())
+	case "tcp":
+		writer, err := newDatadogTCPWriter(addr.String())
+		if err == nil {
+			client, err = statsd.NewWithWriter(writer)
+		}
+	default:
+		err = fmt.Errorf("%s is not supported for sending statsd metrics", network)
+	}
 	if err != nil {
 		return err
 	}
@@ -608,7 +707,7 @@ func sendStatsd(addr string, span *ssf.SSFSpan) error {
 			return err
 		}
 	}
-	client.Flush()
+	client.Close()
 	return nil
 }